% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiply_by_two}
\alias{multiply_by_two}
\title{Multiply a Vector by Two}
\usage{
multiply_by_two(x, n)
}
\arguments{
\item{x}{A \code{numeric} vector of values to multiply.}

\item{n}{The number of elements of \code{x} to process.}
}
\description{
Doubles each element of a numeric vector by passing the data into a
parallelized \emph{C} routine.
}
\examples{

# Data
x = 1:5

# Process all elements of x
n = length(x)

# Call the routine
multiply_by_two(x, n = n)
}
