% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppCandOpenMP-package.R
\docType{package}
\name{RcppCandOpenMP-package}
\alias{RcppCandOpenMP}
\alias{RcppCandOpenMP-package}
\title{RcppCandOpenMP: Using OpenMP with a C Routine and Rcpp}
\description{
Provides an implementation of incorporating parallelization using 'OpenMP' with 'C' and exporting it into 'R' using 'Rcpp'.
}
\details{
Disclaimer. Work contained within package is part of an R&D effort to
show a minimal example of how something can be achieved. Care should be
used when transferring it into production.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless-rd-rcpp/rcpp-c-and-openmp}
  \item \url{https://rd-rcpp.thecoatlessprofessor.com/rcpp-c-and-openmp/}
  \item Report bugs at \url{https://github.com/coatless-rd-rcpp/rcpp-c-and-openmp/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
