% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppOpenMP-package.R
\docType{package}
\name{RcppOpenMP-package}
\alias{RcppOpenMP}
\alias{RcppOpenMP-package}
\title{RcppOpenMP: Using OpenMP with Rcpp}
\description{
Provides an implementation of incorporating parallelization using 'OpenMP'
    with 'Rcpp'.
}
\details{
Disclaimer. Work contained within package is part of an R&D effort to
show a minimal example of how something can be achieved. Care should be
used when transferring it into production.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/r-pkg-examples/rcpp-and-openmp}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
