% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiply_by_two}
\alias{multiply_by_two}
\title{Multiple a Value By 2}
\usage{
multiply_by_two(x, n)
}
\arguments{
\item{x}{A \code{vector} of numeric values to multiple.}

\item{n}{The number of cores to use.}
}
\description{
Computes the multiplication of 2 by passing the data into a \emph{C} routine
that is parallelized
}
\examples{

# Data
x = 1:5

# Number of Cores to use in Parallelization
ncores = 2

# Call the routine
multiply_by_two(x, n = ncores)
}
